+++ rgwin_graph.c
@@ -94,13 +94,13 @@ void win_graphics_rearrange(window_t *win, grect_t *box, data_metrics_t *metrics
     window_graphics_t *dwin = win->data;
     dwin->owner->bbox = *box;
 
-    int width = box->right - box->left;
-    int height = box->bottom - box->top;
+    double width = box->right - box->left;
+    double height = box->bottom - box->top;
 
-    dwin->graphwidth = width - metrics->graphicsmarginx;
+    dwin->graphwidth = (int) (width - metrics->graphicsmarginx);
     if (dwin->graphwidth < 0)
         dwin->graphwidth = 0;
-    dwin->graphheight = height - metrics->graphicsmarginy;
+    dwin->graphheight = (int) (height - metrics->graphicsmarginy);
     if (dwin->graphheight < 0)
         dwin->graphheight = 0;
 }
